refactor: move whatsapp and calengrade components to feature folders (#510)
